auth_app/pyproject.toml
[project]
name = "auth-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan",email = "iskiba60@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
